add design for Additional commands
