--- cmd/diagnose.go
@@ -1,40 +1,0 @@
-/*
-Copyright © 2024 NAME HERE <EMAIL ADDRESS>
-
-*/
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-// diagnoseCmd represents the diagnose command
-var diagnoseCmd = &cobra.Command{
-	Use:   "diagnose",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("diagnose called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(diagnoseCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// diagnoseCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// diagnoseCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
